[Feature implementation] Implement Ktor Plugin to the core library. (#34)
* Divide the Configuration class.
